Rename collector case generator module
