+++ collector/sglang/collect_moe.py
@@ -84,7 +84,7 @@ except ImportError:
     pass
 
 try:
-    from case_specs import get_common_moe_test_cases
+    from case_generator import get_common_moe_test_cases
 
     from helper import (
         balanced_logits,
@@ -99,7 +99,7 @@ except ModuleNotFoundError:
     import sys
 
     sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-    from case_specs import get_common_moe_test_cases
+    from case_generator import get_common_moe_test_cases
 
     from helper import (
         balanced_logits,
