+++ collector/trtllm/collect_mamba2.py
@@ -46,7 +46,7 @@ import torch
 from einops import repeat
 
 try:
-    from case_specs import get_common_mamba2_test_cases
+    from case_generator import get_common_mamba2_test_cases
 
     from helper import (
         EXIT_CODE_RESTART,
@@ -58,7 +58,7 @@ except ModuleNotFoundError:
     import sys
 
     sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-    from case_specs import get_common_mamba2_test_cases
+    from case_generator import get_common_mamba2_test_cases
 
     from helper import (
         EXIT_CODE_RESTART,
