+++ collector/trtllm/collect_moe_v1.py
@@ -19,7 +19,7 @@ from tensorrt_llm.mapping import Mapping
 from tensorrt_llm.models.modeling_utils import QuantAlgo, QuantConfig
 from torch.nn.parameter import Parameter
 
-from collector.case_specs import get_common_moe_test_cases
+from collector.case_generator import get_common_moe_test_cases
 from collector.helper import (
     balanced_logits,
     benchmark_with_power,
