+++ collector/vllm/collect_moe_v2.py
@@ -86,7 +86,7 @@ except Exception:
 
 from vllm.forward_context import get_forward_context, set_forward_context
 
-from collector.case_specs import get_common_moe_test_cases
+from collector.case_generator import get_common_moe_test_cases
 from collector.helper import balanced_logits, benchmark_with_power, get_sm_version, log_perf, power_law_logits_v3
 
 aic_debug = int(os.getenv("aic_moe_debug", "0"))  # noqa: SIM112
