+++ collector/wideep/trtllm/collect_moe_compute.py
@@ -43,7 +43,7 @@ if COLLECTOR_ROOT not in sys.path:
     sys.path.append(COLLECTOR_ROOT)
 
 try:
-    from case_specs import get_common_moe_test_cases, is_wideep_moe_model
+    from case_generator import get_common_moe_test_cases, is_wideep_moe_model
 
     from helper import (
         EXIT_CODE_RESTART,
@@ -55,7 +55,7 @@ try:
     )
 except ModuleNotFoundError:
     sys.path.append(COLLECTOR_ROOT)
-    from case_specs import get_common_moe_test_cases, is_wideep_moe_model
+    from case_generator import get_common_moe_test_cases, is_wideep_moe_model
 
     from helper import (
         EXIT_CODE_RESTART,
