+++ tests/unit/sdk/database/test_dsv4_flash_sparse.py
@@ -499,7 +499,7 @@ def test_lookup_sparse_kernel_missing_returns_none():
 
 def test_dsv4_flash_test_cases_active_under_no_filter(monkeypatch):
     monkeypatch.delenv("COLLECTOR_MODEL_PATH", raising=False)
-    from collector.case_specs import (
+    from collector.case_generator import (
         get_dsv4_flash_csa_context_test_cases,
         get_dsv4_flash_paged_mqa_logits_test_cases,
     )
@@ -511,7 +511,7 @@ def test_dsv4_flash_test_cases_active_under_no_filter(monkeypatch):
 def test_dsv4_flash_test_cases_skipped_under_other_model(monkeypatch):
     """Filter to a non-V4 model → V4 ops emit zero cases (collector skips)."""
     monkeypatch.setenv("COLLECTOR_MODEL_PATH", "deepseek-ai/DeepSeek-V3")
-    from collector.case_specs import (
+    from collector.case_generator import (
         get_dsv4_flash_csa_context_test_cases,
         get_dsv4_flash_csa_generation_test_cases,
         get_dsv4_flash_hca_attn_test_cases,
@@ -527,7 +527,7 @@ def test_dsv4_flash_test_cases_skipped_under_other_model(monkeypatch):
 def test_dsv4_flash_test_cases_active_under_v4_filter(monkeypatch):
     model_path = "sgl-project/DeepSeek-V4-Flash-FP8"
     monkeypatch.setenv("COLLECTOR_MODEL_PATH", model_path)
-    from collector.case_specs import get_dsv4_flash_csa_context_test_cases
+    from collector.case_generator import get_dsv4_flash_csa_context_test_cases
 
     cases = get_dsv4_flash_csa_context_test_cases()
     assert len(cases) > 0
@@ -540,7 +540,7 @@ def test_dsv4_flash_test_cases_active_under_v4_filter(monkeypatch):
 def test_dsv4_flash_sparse_test_cases_only_indexer_tp1(monkeypatch):
     """Sweep is fixed at tp=[1] (kernel is TP-invariant)."""
     monkeypatch.setenv("COLLECTOR_MODEL_PATH", "sgl-project/DeepSeek-V4-Flash-FP8")
-    from collector.case_specs import (
+    from collector.case_generator import (
         get_dsv4_flash_hca_attn_test_cases,
         get_dsv4_flash_paged_mqa_logits_test_cases,
     )
